requirements.txt
fastapi
uvicorn[standard]
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
python-dotenv
pytest
email-validator
pydantic
dateparser
